# srs/openapi3_0.yaml
openapi: 3.0.3
info:
  title: Authentication and Role Management API
  description: API для аутентификации и управления ролями пользователей.
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
    description: Основной сервер
paths:
  /auth/register:
    post:
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
        '400':
          description: Ошибка валидации данных
...
  /auth/me:
    get:
      summary: Личный кабинет
      responses:
        '200':
          description: Личный кабинет пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  password:
                    type: string
        '401':
          description: Неавторизованный запрос
...